Fix sporadic bug in day 20 in Rust
The check for an already correct orientation of the start corner was the wrong way around. This only sporadically caused errors because which of the four corners was used as the start corner depended the order of iteration on a HashMap, which is random.
The randomness has been removed as well by explicitly using the corner with the smallest id.